Clustering initial commit
