+++ src/commands/game_commands/join.ts
@@ -6,7 +6,7 @@ import { GuildTextableMessage, ParsedMessage, GameType } from "../../types";
 import { getUserTag, sendErrorMessage, sendInfoMessage } from "../../helpers/discord_utils";
 import { KmqImages } from "../../constants";
 import { bold } from "../../helpers/utils";
-import state from "../../kmq";
+import { state } from "../../kmq";
 import MessageContext from "../../structures/message_context";
 import KmqMember from "../../structures/kmq_member";
