+++ src/events/client/voiceChannelLeave.ts
@@ -1,5 +1,5 @@
 import Eris from "eris";
-import state from "../../kmq";
+import { state } from "../../kmq";
 import { checkBotIsAlone } from "../../helpers/discord_utils";
 
 export default async function voiceChannelLeaveHandler(member: Eris.Member, oldChannel: Eris.VoiceChannel) {
