+++ src/helpers/game_utils.ts
@@ -1,5 +1,5 @@
 import dbContext from "../database_context";
-import state from "../kmq";
+import { state } from "../kmq";
 import _logger from "../logger";
 import GameSession from "../structures/game_session";
 import GuildPreference from "../structures/guild_preference";
