+++ src/structures/guild_preference.ts
@@ -14,7 +14,7 @@ import { MatchedArtist } from "../types";
 import { DEFAULT_OST_PREFERENCE, OstPreference } from "../commands/game_options/ost";
 import { DEFAULT_RELEASE_TYPE, ReleaseType } from "../commands/game_options/release";
 import { DEFAULT_MULTIGUESS_TYPE, MultiGuessType } from "../commands/game_options/multiguess";
-import state from "../kmq";
+import { state } from "../kmq";
 import { SpecialType } from "../commands/game_options/special";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
