+++ src/structures/message_context.ts
@@ -1,6 +1,6 @@
 import Eris from "eris";
 import { getUserTag } from "../helpers/discord_utils";
-import state from "../kmq";
+import { state } from "../kmq";
 import KmqMember from "./kmq_member";
 
 export default class MessageContext {
