+++ src/test/ci/game_utils.test.ts
@@ -12,7 +12,7 @@ import _logger from "../../logger";
 import { ArtistType } from "../../commands/game_options/artisttype";
 import { SubunitsPreference } from "../../commands/game_options/subunits";
 import { FOREIGN_LANGUAGE_TAGS, LanguageType } from "../../commands/game_options/language";
-import state from "../../kmq";
+import { state } from "../../kmq";
 import GameSession from "../../structures/game_session";
 import { OstPreference } from "../../commands/game_options/ost";
 import { NON_OFFICIAL_VIDEO_TAGS, ReleaseType } from "../../commands/game_options/release";
