+++ start.sh
@@ -11,10 +11,10 @@ node build/seed/bootstrap.js
 echo "Starting bot..."
 cd build/
 if [ "${NODE_ENV}" == "dry-run" ] || [ "${NODE_ENV}" == "ci" ]; then
-    exec node kmq.js
+    exec node index.js
 elif [ "${NODE_ENV}" == "development" ]; then
-    exec node --inspect=9229 kmq.js
+    exec node --inspect=9229 index.js
 elif [ "${NODE_ENV}" == "production" ]; then
     git log -n 1 --pretty=format:"%H" > ../version
-    exec node kmq.js
+    exec node index.js
 fi
